fix(admin): apply code-review findings
- split RevelUserAdmin filter + support inlines into
  accounts/admin/user_extras.py (user.py was over the 1000-line limit)
- make Payment base amounts (amount, currency, platform_fee) readonly
  so admin edits can't diverge from Stripe (CodeRabbit)
- lock NotificationDeliveryAdmin status/retry_count/error_message to
  match the read-only inline (CodeRabbit)
- add de/fr/it translations for the new "organization owner" filter
