+++ src/events/admin/ticket.py
@@ -170,6 +170,9 @@ class PaymentAdmin(ModelAdmin, UserLinkMixin, EventLinkMixin):  # type: ignore[m
         "id",
         "user",
         "ticket",
+        "amount",
+        "currency",
+        "platform_fee",
         "stripe_session_id",
         "stripe_payment_intent_id",
         "reservation_id",
