+++ src/notifications/admin.py
@@ -283,6 +283,9 @@ class NotificationDeliveryAdmin(admin.ModelAdmin):  # type: ignore[type-arg]
         "id",
         "notification",
         "channel",
+        "status",
+        "retry_count",
+        "error_message",
         "attempted_at",
         "delivered_at",
         "created_at",
